Change lifx protocol
The lifxlan library was no longer working either was the photon library. Basically the lan protocol stopped working in January. For some reason homekit was not detecting them either.

So to alleviate the problem we have switched to the http api protocol. This is a little slower, but none the less does the trick.